frex v0.0.1 Frex.Client.Categories

Client module for interacting with the Freshbooks categories endpoint.

Staff have access to the get, list and create API calls listed below if the Expenses tab is enabled for staff. Staff cannot delete or update Categories.

Functions

get(credentials, category_id)

Return the complete category details associated with the given category_id.

Parameters

category_id (required) — categories ID

list(credentials, filters \\ %{})

Returns a list of expense categories.