frex v0.1.3 Frex.Client.Gateways

Client module for interacting with the Freshbooks gateways endpoint.

The Gateway API provides access to the Online Payment Gateways enabled in the application.

Functions

list(credentials, filters \\ %{})

Returns a list of payment gateways enabled in your FreshBooks account that can process transactions. You can optionally filter by autobill_capable to return only gateways that support auto-bills.

Parameters

filters (optional) — a map of filters for the list request

autobill_capable (boolean, 1 = true, 0 = false)