frex v0.0.1 Frex.Client.Reports

Provides access to the most commonly used reports and information.

Functions

get_client_details(credentials, client_id, filters \\ %{})

Returns profit and account standing details for the specified client.

Parameters

client_id

get_estimate_details(credentials, client_id, filters \\ %{})

Returns total estimated, accepted and invoiced values for invoices; broken down by currency.

Parameters

client_id

get_expense_details(credentials, filters \\ %{})

Returns information about all, or a filtered set of expenses.

Parameters

client_id (optional)

get_invoice_details(credentials, filters \\ %{})

Returns information about all, or a filtered set of invoices. Default sorting in the response is by currency code, unless overruled with the filter.

Parameters

client_id (optional)

get_profit_details(credentials, date_from, date_to, filters \\ %{})

Returns profit & loss details. With this report we assume that the currency_code is the same as the system’s base currency_code. Invoice amounts in other currencies will be converted. The normal call (not filtered by date) will return the same information as the ‘Totals’ column on the Profit & Loss Report.

Parameters

date_from YYYY-MM-DD format

date_to YYYY-MM-DD format

group_by (optional)