frex v0.0.1 Frex.Client.TimeEntries

Client module for interacting with the Freshbooks time entries endpoint.

Functions

get(credentials, time_entry_id)

Sends a get request to the Freshbooks API timeentries endpoint.

Parameters

time_entry_id (required) — time_entry ID

list(credentials, filters \\ %{})

Returns a list of timesheet entries ordered according to date.

Parameters

filters (optional) — a map of filters for the list request

project_id

task_id

date_from

date_to