frex v0.1.0 Frex.Request

Handles the building of requests for sending to the Freshbooks API.

Functions

build(method)

Builds an XML request from a given method and XML list.

Examples

Frex.Request.build("expenses.get", {:expense_id, %{}, 433})

# => "<?xml version="1.0" encoding="UTF-8" ?>
<request method="expenses.get">
    <expense_id>433</expense_id>
</request>"

build(method, builder_list)

list_all(creds, list_fun, acc \\ [], inc \\ 1)

Paginates through a list to get all data in one big list.

Returns a list.

more_pages?(arg1)

Checks the current page data to see if more pages need to be loaded.

Returns a boolean.